Move compose command logic to a separate file
